refactor pcap reading
